+++ arcade_machine/games/arcade_settings.py
@@ -1,11 +1,9 @@
 
 from arcade_machine.games.game import Game
 from arcade_machine.sprites.label import Label
-from arcade_machine.components.music_player import music_volume, get_volume
-from arcade_machine.components.shape_to_sprite import rect_sprite
+from arcade_machine.Controller.music_player import music_volume, get_volume
 from arcade_machine.events import CHANGE_GAME
 from pygame import font
-from pygame import draw
 from pygame import KEYDOWN, K_a, K_d, K_j, K_l, K_1, K_9, K_m
 from pygame import quit as pygame_quit
 from pygame.event import post as pygame_post_event
@@ -21,19 +19,21 @@ class ArcadeSettings(Game):
 
         self.TITLE_FONT = font.Font('arcade_machine/resources/fonts/Main Menu/Early GameBoy.ttf', 48)
         self.BODY_FONT = font.Font('arcade_machine/resources/fonts/Main Menu/Early GameBoy.ttf', 24)
-        self.menu_title = Label('Settings', (255, 255, 255), 512, 60, None, self.TITLE_FONT)
-        self.drawable_objects.append(self.menu_title)
+
+        self.settings_title = Label('Settings', (255, 255, 255), 512, 60, None, self.TITLE_FONT)
         self.vol_text = Label('System Volume:', (255, 255, 255), 512, 200, None, self.BODY_FONT)
-        self.drawable_objects.append(self.vol_text)
         self.vol_level = Label(str(int(self.volume * 100)), (255, 255, 255), 512, 250, None, self.BODY_FONT)
-        self.drawable_objects.append(self.vol_level)
         self.ver_text = Label('Version 0.0.1', (100, 100, 100), 512, 400, None, self.BODY_FONT)
-        self.drawable_objects.append(self.ver_text)
         self.back = Label("Press 'A' to return to the games.", (120, 120, 120), 512, 530, None, self.BODY_FONT)
-        self.drawable_objects.append(self.back)
         self.leave = Label('Press MENU to quit the console.', (80, 10, 15), 512, 575, None, self.BODY_FONT)
-        self.drawable_objects.append(self.leave)
         self.creators = Label('By Alex and Billy', (40, 40, 40), 512, 670, None, self.BODY_FONT)
+
+        self.drawable_objects.append(self.settings_title) # Add labels to be drawn
+        self.drawable_objects.append(self.vol_text)
+        self.drawable_objects.append(self.vol_level)
+        self.drawable_objects.append(self.ver_text)
+        self.drawable_objects.append(self.back)
+        self.drawable_objects.append(self.leave)
         self.drawable_objects.append(self.creators)
 
         #self.line = rect_sprite((100, 0, 1024, 12))
@@ -52,15 +52,15 @@ class ArcadeSettings(Game):
                 pygame_quit()
                 exit()
     def update(self):
-        if self.volume > 1.0:
+        if self.volume > 1.0: # Ensure the volume level is in the pygame mixer limits
             self.volume = 1.0
         elif self.volume < 0:
             self.volume = 0
-        music_volume(self.volume)
+        music_volume(self.volume) # Set the pygame mixer volume level
 
-        self.drawable_objects.clear()
+        self.drawable_objects.clear() # Remove sprites and prepare to draw the next frame
 
-        self.drawable_objects.append(self.menu_title)
+        self.drawable_objects.append(self.settings_title)
         self.drawable_objects.append(self.vol_text)
         self.vol_level = Label(str(int(self.volume * 100)), (255, 255, 255), 512, 250, None, self.BODY_FONT)
         self.drawable_objects.append(self.vol_level)